+++ src/test/java/com/github/fppt/jedismock/comparisontests/sets/SetOperationsTest.java
@@ -113,159 +113,6 @@ public class SetOperationsTest {
         assertFalse(jedis.sismember(key + "-nonexistent", "A"));
     }
 
-    @TestTemplate
-    public void whenUsingHsinter_EnsureSetIntersectionIsReturned(Jedis jedis) {
-        String key1 = "my-set-key-1";
-        Set<String> mySet1 = new HashSet<>(Arrays.asList("a", "b", "c", "d"));
-        String key2 = "my-set-key-2";
-        Set<String> mySet2 = new HashSet<>(Arrays.asList("b", "d", "e", "f"));
-        String key3 = "my-set-key-3";
-        Set<String> mySet3 = new HashSet<>(Arrays.asList("b", "e", "f"));
-
-        Set<String> expectedIntersection1 = new HashSet<>(Arrays.asList("b", "d"));
-        Set<String> expectedIntersection2 = new HashSet<>(Collections.singletonList("b"));
-
-        //Add everything from the sets
-        mySet1.forEach(value -> jedis.sadd(key1, value));
-        mySet2.forEach(value -> jedis.sadd(key2, value));
-        mySet3.forEach(value -> jedis.sadd(key3, value));
-
-        Set<String> intersection = jedis.sinter(key1, key2);
-        assertEquals(expectedIntersection1, intersection);
-
-        intersection = jedis.sinter(key1, key2, key3);
-        assertEquals(expectedIntersection2, intersection);
-    }
-
-    @TestTemplate
-    public void whenUsingSInterStore_testIntersectionIsStorred(Jedis jedis) {
-        String key1 = "set1";
-        String key2 = "set2";
-        Set<String> mySet1 = new HashSet<>(Arrays.asList("a", "b", "c", "d"));
-        Set<String> mySet2 = new HashSet<>(Arrays.asList("b", "d", "e", "f"));
-
-        Set<String> expectedIntersection = new HashSet<>(Arrays.asList("b", "d"));
-
-        //Add everything from the sets
-        mySet1.forEach(value -> jedis.sadd(key1, value));
-        mySet2.forEach(value -> jedis.sadd(key2, value));
-
-        String destination = "set3";
-
-        Long elementsInIntersection = jedis.sinterstore(destination, key1, key2);
-        assertEquals(2, elementsInIntersection);
-
-        assertEquals(expectedIntersection, jedis.smembers(destination));
-    }
-
-    @TestTemplate
-    public void sUnionTest(Jedis jedis) {
-        String key1 = "set1";
-        String key2 = "set2";
-        String key3 = "set3";
-        Set<String> mySet1 = new HashSet<>(Arrays.asList("a", "b", "c", "d"));
-        Set<String> mySet2 = new HashSet<>(Arrays.asList("c"));
-        Set<String> mySet3 = new HashSet<>(Arrays.asList("a", "c", "e", "f"));
-
-        Set<String> expectedUnion = new HashSet<>(Arrays.asList("a", "b", "c", "d", "e", "f"));
-
-        //Add everything from the sets
-        mySet1.forEach(value -> jedis.sadd(key1, value));
-        mySet2.forEach(value -> jedis.sadd(key2, value));
-        mySet3.forEach(value -> jedis.sadd(key3, value));
-
-
-        Set<String> result = jedis.sunion(key1, key2, key3);
-        assertEquals(6, result.size());
-        assertEquals(expectedUnion, result);
-    }
-
-    @TestTemplate
-    public void sUnionStoreTest(Jedis jedis) {
-        String key1 = "set1";
-        String key2 = "set2";
-        String key3 = "set3";
-        Set<String> mySet1 = new HashSet<>(Arrays.asList("a", "b", "c", "d"));
-        Set<String> mySet2 = new HashSet<>(Arrays.asList("c"));
-        Set<String> mySet3 = new HashSet<>(Arrays.asList("a", "c", "e", "f"));
-
-        Set<String> expectedUnion = new HashSet<>(Arrays.asList("a", "b", "c", "d", "e", "f"));
-
-        //Add everything from the sets
-        mySet1.forEach(value -> jedis.sadd(key1, value));
-        mySet2.forEach(value -> jedis.sadd(key2, value));
-        mySet3.forEach(value -> jedis.sadd(key3, value));
-
-        String destination = "set3";
-
-        Long elementsInUnion = jedis.sunionstore(destination, key1, key2, key3);
-        assertEquals(6, elementsInUnion);
-
-        assertEquals(expectedUnion, jedis.smembers(destination));
-    }
-
-    @TestTemplate
-    public void sDiffTwoSetsTest(Jedis jedis) {
-        String key1 = "set1";
-        String key2 = "set2";
-        Set<String> mySet1 = new HashSet<>(Arrays.asList("a", "b", "c", "d"));
-        Set<String> mySet2 = new HashSet<>(Arrays.asList("c", "d", "e"));
-
-        Set<String> expectedDifference = new HashSet<>(Arrays.asList("a", "b"));
-
-        //Add everything from the sets
-        mySet1.forEach(value -> jedis.sadd(key1, value));
-        mySet2.forEach(value -> jedis.sadd(key2, value));
-
-
-        Set<String> result = jedis.sdiff(key1, key2);
-        assertEquals(2, result.size());
-        assertEquals(expectedDifference, result);
-    }
-
-    @TestTemplate
-    public void sDiffThreeSetsTest(Jedis jedis) {
-        String key1 = "set1";
-        String key2 = "set2";
-        String key3 = "set3";
-        Set<String> mySet1 = new HashSet<>(Arrays.asList("a", "b", "c", "d"));
-        Set<String> mySet2 = new HashSet<>(Arrays.asList("c"));
-        Set<String> mySet3 = new HashSet<>(Arrays.asList("a", "c", "e"));
-
-        Set<String> expectedDifference = new HashSet<>(Arrays.asList("b", "d"));
-
-        //Add everything from the sets
-        mySet1.forEach(value -> jedis.sadd(key1, value));
-        mySet2.forEach(value -> jedis.sadd(key2, value));
-        mySet3.forEach(value -> jedis.sadd(key3, value));
-
-
-        Set<String> result = jedis.sdiff(key1, key2, key3);
-        assertEquals(2, result.size());
-        assertEquals(expectedDifference, result);
-    }
-
-    @TestTemplate
-    public void sDiffStoreTest(Jedis jedis) {
-        String key1 = "set1";
-        String key2 = "set2";
-        Set<String> mySet1 = new HashSet<>(Arrays.asList("a", "b", "c", "d"));
-        Set<String> mySet2 = new HashSet<>(Arrays.asList("c", "d", "e"));
-
-        Set<String> expectedDifference = new HashSet<>(Arrays.asList("a", "b"));
-
-        //Add everything from the sets
-        mySet1.forEach(value -> jedis.sadd(key1, value));
-        mySet2.forEach(value -> jedis.sadd(key2, value));
-
-        String destination = "set3";
-
-        Long elementsInDifference = jedis.sdiffstore(destination, key1, key2);
-        assertEquals(2, elementsInDifference);
-
-        assertEquals(expectedDifference, jedis.smembers(destination));
-    }
-    
 
     @TestTemplate
     public void testFailingGetOperation(Jedis jedis) {
